requirements.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
asgiref==3.3.1            # via django
django==3.1.5
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2020.5              # via django
sqlparse==0.4.1           # via django

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via gunicorn
